refactor: jwt
