+++ backend/Core/Boundaries/Jwt/JwtTokenProvider.cs
@@ -4,7 +4,7 @@ using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
 
-namespace Core.Boundaries;
+namespace Core.Boundaries.Jwt;
 
 public class JwtTokenProvider(IConfiguration configuration)
 {
